requirements.txt
opencv-python
numpy
PyPDF2
Pillow
pytesseract
pdf2image